requirements.txt
flask
azure-cognitiveservices-vision-computervision
msrest
psycopg2-binary
python-dotenv
gunicorn